LayoutTests/http/tests/security/block-top-level-navigations-by-third-party-iframe-sandboxed-by-own-csp-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to initiate navigation for frame with URL 'http://127.0.0.1:8000/security/block-top-level-navigations-by-third-party-iframe-sandboxed-by-own-csp.html' from frame with URL 'http://127.0.0.1:8000/security/resources/attempt-top-level-navigation-with-csp.py'. The frame attempting navigation of the top-level window is cross-origin or untrusted and the user has never interacted with the frame.

CONSOLE MESSAGE: SecurityError: The operation is insecure.
Test blocking of top-level navigations by a third-party iframe which gives itself a sandbox which allows top navigation via CSP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS All navigations by subframes have been blocked
PASS successfullyParsed is true

TEST COMPLETE
